FarmLand and fixes and new commands (more to come -.-)
