+++ src/main/java/org/zone/commands/structure/ZoneCommands.java
@@ -8,14 +8,16 @@ import org.zone.commands.structure.create.bounds.exact.ZoneCreateStartCommand;
 import org.zone.commands.structure.create.bounds.exact.ZoneCreateSubStartCommand;
 import org.zone.commands.structure.info.ZonePluginInfoCommand;
 import org.zone.commands.structure.leave.LeaveZoneCommand;
-import org.zone.commands.structure.region.flags.damage.attack.ZoneFlagDamageAttackSetEnabledCommand;
-import org.zone.commands.structure.region.flags.damage.attack.ZoneFlagDamageAttackSetGroupCommand;
-import org.zone.commands.structure.region.flags.damage.attack.ZoneFlagDamageAttackView;
-import org.zone.commands.structure.region.flags.damage.fall.ZoneFlagPlayerFallDamageEnableDisable;
-import org.zone.commands.structure.region.flags.damage.fall.ZoneFlagPlayerFallDamageSetGroup;
-import org.zone.commands.structure.region.flags.damage.fall.ZoneFlagPlayerFallDamageView;
-import org.zone.commands.structure.region.flags.damage.tnt.ZoneFlagTntDefuseSetEnableDisableCommand;
-import org.zone.commands.structure.region.flags.damage.tnt.ZoneFlagTntDefuseView;
+import org.zone.commands.structure.region.flags.damage.to.block.farmland.ZoneFlagFarmLandTrampleEnableCommand;
+import org.zone.commands.structure.region.flags.damage.to.block.farmland.ZoneFlagFarmLandTrampleViewCommand;
+import org.zone.commands.structure.region.flags.damage.to.player.attack.ZoneFlagDamageAttackSetEnabledCommand;
+import org.zone.commands.structure.region.flags.damage.to.player.attack.ZoneFlagDamageAttackSetGroupCommand;
+import org.zone.commands.structure.region.flags.damage.to.player.attack.ZoneFlagDamageAttackView;
+import org.zone.commands.structure.region.flags.damage.to.player.fall.ZoneFlagPlayerFallDamageEnableDisable;
+import org.zone.commands.structure.region.flags.damage.to.player.fall.ZoneFlagPlayerFallDamageSetGroup;
+import org.zone.commands.structure.region.flags.damage.to.player.fall.ZoneFlagPlayerFallDamageView;
+import org.zone.commands.structure.region.flags.damage.to.both.tnt.ZoneFlagTntDefuseSetEnableDisableCommand;
+import org.zone.commands.structure.region.flags.damage.to.both.tnt.ZoneFlagTntDefuseView;
 import org.zone.commands.structure.region.flags.eco.ZoneFlagViewBalanceCommand;
 import org.zone.commands.structure.region.flags.entry.monster.ZoneFlagMonsterEntryEnabledCommand;
 import org.zone.commands.structure.region.flags.entry.monster.ZoneFlagMonsterEntryViewCommand;
@@ -81,6 +83,8 @@ public interface ZoneCommands {
     ZoneFlagGreetingsViewCommand ZONE_FLAG_GREETINGS_VIEW_COMMAND = new ZoneFlagGreetingsViewCommand();
     ZoneFlagInteractItemframesViewCommand ZONE_FLAG_INTERACT_ITEMFRAMES_VIEW_COMMAND = new ZoneFlagInteractItemframesViewCommand();
 
+    ZoneFlagFarmLandTrampleEnableCommand ZONE_FLAG_FARM_LAND_TRAMPLE_ENABLE_COMMAND = new ZoneFlagFarmLandTrampleEnableCommand();
+    ZoneFlagFarmLandTrampleViewCommand ZONE_FLAG_FARM_LAND_TRAMPLE_VIEW_COMMAND = new ZoneFlagFarmLandTrampleViewCommand();
     ZoneFlagTntDefuseSetEnableDisableCommand ZONE_FLAG_TNT_DEFUSE_SET_ENABLE_DISABLE_COMMAND = new ZoneFlagTntDefuseSetEnableDisableCommand();
     ZoneFlagTntDefuseView ZONE_FLAG_TNT_DEFUSE_VIEW = new ZoneFlagTntDefuseView();
     ZoneFlagPlayerFallDamageSetGroup ZONE_FLAG_PLAYER_FALL_DAMAGE_SET_GROUP = new ZoneFlagPlayerFallDamageSetGroup();
