+++ src/main/java/org/zone/commands/structure/join/invite/InvitePlayerCommand.java
@@ -1,10 +1,9 @@
-package org.zone.commands.structure.join;
+package org.zone.commands.structure.join.invite;
 
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.configurate.ConfigurateException;
 import org.zone.commands.system.ArgumentCommand;
 import org.zone.commands.system.CommandArgument;
 import org.zone.commands.system.arguments.operation.ExactArgument;
@@ -12,27 +11,24 @@ import org.zone.commands.system.arguments.zone.ZoneArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.permissions.ZonePermission;
 import org.zone.region.Zone;
-import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.meta.request.join.JoinRequestFlag;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class JoinZoneCommand implements ArgumentCommand {
+public class InvitePlayerCommand implements ArgumentCommand {
 
-    public static final ZoneArgument ZONE_ID = new ZoneArgument("zoneId",
-            new ZoneArgument.ZoneArgumentPropertiesBuilder().setVisitorOnly(true));
+    public static final ZoneArgument ZONE_ID = new ZoneArgument("zoneId", new ZoneArgument.ZoneArgumentPropertiesBuilder().setVisitorOnly(false));
 
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
-        return Arrays.asList(new ExactArgument("join"), ZONE_ID);
+        return Arrays.asList(new ExactArgument("invite"), ZONE_ID);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("Join a zone!");
+        return Component.text("Invite a player to a zone");
     }
 
     @Override
@@ -47,18 +43,6 @@ public class JoinZoneCommand implements ArgumentCommand {
             return CommandResult.error(Messages.getPlayerOnlyMessage());
         }
         Zone zone = commandContext.getArgument(this, ZONE_ID);
-        JoinRequestFlag joinRequestFlag = zone
-                .getFlag(FlagTypes.JOIN_REQUEST)
-                .orElse(new JoinRequestFlag());
-        joinRequestFlag.registerInvite(player.uniqueId());
-        zone.setFlag(joinRequestFlag);
-        try {
-            zone.save();
-            commandContext.sendMessage(Messages.getJoinedZoneMessage(zone));
-        } catch (ConfigurateException ce) {
-            ce.printStackTrace();
-            return CommandResult.error(Messages.getZoneSavingError(ce));
-        }
         return CommandResult.success();
     }
 }
