+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/both/tnt/ZoneFlagTntDefuseView.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.damage.tnt;
+package org.zone.commands.structure.region.flags.damage.to.both.tnt;
 
 import net.kyori.adventure.identity.Identity;
 import net.kyori.adventure.text.Component;
