+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/player/attack/ZoneFlagDamageAttackSetEnabledCommand.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.damage.fall;
+package org.zone.commands.structure.region.flags.damage.to.player.attack;
 
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
@@ -7,27 +7,28 @@ import org.spongepowered.configurate.ConfigurateException;
 import org.zone.commands.system.ArgumentCommand;
 import org.zone.commands.system.CommandArgument;
 import org.zone.commands.system.arguments.operation.ExactArgument;
+import org.zone.commands.system.arguments.simple.BooleanArgument;
 import org.zone.commands.system.arguments.zone.ZoneArgument;
-import org.zone.commands.system.arguments.zone.ZoneGroupArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.permissions.ZonePermission;
 import org.zone.permissions.ZonePermissions;
 import org.zone.region.Zone;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.entity.player.damage.fall.PlayerFallDamageFlag;
-import org.zone.region.group.Group;
+import org.zone.region.flag.entity.player.damage.attack.EntityDamagePlayerFlag;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class ZoneFlagPlayerFallDamageSetGroup implements ArgumentCommand {
+public class ZoneFlagDamageAttackSetEnabledCommand implements ArgumentCommand {
 
     public static final ZoneArgument ZONE_VALUE = new ZoneArgument("zoneId",
             new ZoneArgument.ZoneArgumentPropertiesBuilder().setBypassSuggestionPermission(
-                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_FALL_SET_GROUP));
-    public static final ZoneGroupArgument GROUP = new ZoneGroupArgument("groupId", ZONE_VALUE);
+                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_ATTACK_ENABLE));
+    public static final BooleanArgument ENABLED = new BooleanArgument("enableValue",
+            "enable",
+            "disable");
 
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
@@ -35,35 +36,41 @@ public class ZoneFlagPlayerFallDamageSetGroup implements ArgumentCommand {
                 new ExactArgument("flag"),
                 ZONE_VALUE,
                 new ExactArgument("damage"),
-                new ExactArgument("fall"),
+                new ExactArgument("attack"),
+                new ExactArgument("towards"),
+                new ExactArgument("player"),
                 new ExactArgument("set"),
-                new ExactArgument("group"),
-                GROUP);
+                ENABLED);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("Sets the group for Entity Damage Player flag");
+        return Component.text("Command to enable/disable the Damage Flag");
     }
 
     @Override
     public @NotNull Optional<ZonePermission> getPermissionNode() {
-        return Optional.of(ZonePermissions.FLAG_DAMAGE_FALL_SET_GROUP);
+        return Optional.of(ZonePermissions.FLAG_DAMAGE_ATTACK_ENABLE);
     }
 
     @Override
     public @NotNull CommandResult run(
             @NotNull CommandContext commandContext, @NotNull String... args) {
+        boolean enable = commandContext.getArgument(this, ENABLED);
         Zone zone = commandContext.getArgument(this, ZONE_VALUE);
-        PlayerFallDamageFlag playerFallDamageFlag = zone
-                .getFlag(FlagTypes.PLAYER_FALL_DAMAGE)
-                .orElse(FlagTypes.PLAYER_FALL_DAMAGE.createCopyOfDefault());
-        Group group = commandContext.getArgument(this, GROUP);
-        zone.getMembers().addKey(group, playerFallDamageFlag.getRequiredKey());
+        EntityDamagePlayerFlag entityDamagePlayerFlag = zone
+                .getFlag(FlagTypes.ENTITY_DAMAGE_PLAYER)
+                .orElse(new EntityDamagePlayerFlag());
+        if (enable) {
+            zone.addFlag(entityDamagePlayerFlag);
+        } else {
+            zone.removeFlag(FlagTypes.ENTITY_DAMAGE_PLAYER);
+        }
         try {
             zone.save();
-            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.PLAYER_FALL_DAMAGE));
+            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.ENTITY_DAMAGE_PLAYER));
         } catch (ConfigurateException ce) {
+            ce.printStackTrace();
             return CommandResult.error(Messages.getZoneSavingError(ce));
         }
         return CommandResult.success();
