+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/player/attack/ZoneFlagDamageAttackSetGroupCommand.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.damage.tnt;
+package org.zone.commands.structure.region.flags.damage.to.player.attack;
 
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
@@ -7,68 +7,65 @@ import org.spongepowered.configurate.ConfigurateException;
 import org.zone.commands.system.ArgumentCommand;
 import org.zone.commands.system.CommandArgument;
 import org.zone.commands.system.arguments.operation.ExactArgument;
-import org.zone.commands.system.arguments.simple.BooleanArgument;
 import org.zone.commands.system.arguments.zone.ZoneArgument;
+import org.zone.commands.system.arguments.zone.ZoneGroupArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.permissions.ZonePermission;
 import org.zone.permissions.ZonePermissions;
 import org.zone.region.Zone;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.entity.nonliving.block.tnt.TnTDefuseFlag;
+import org.zone.region.flag.entity.player.damage.attack.EntityDamagePlayerFlag;
+import org.zone.region.group.Group;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class ZoneFlagTntDefuseSetEnableDisableCommand implements ArgumentCommand {
+public class ZoneFlagDamageAttackSetGroupCommand implements ArgumentCommand {
 
     public static final ZoneArgument ZONE_VALUE = new ZoneArgument("zoneId",
             new ZoneArgument.ZoneArgumentPropertiesBuilder().setBypassSuggestionPermission(
-                    ZonePermissions.OVERRIDE_FLAG_TNT_DEFUSE_ENABLE));
-    public static final BooleanArgument ENABLE_DISABLE = new BooleanArgument("enabledValue",
-            "enable",
-            "disable");
+                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_ATTACK_SET_GROUP));
+    public static final ZoneGroupArgument GROUP = new ZoneGroupArgument("groupId", ZONE_VALUE);
 
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
         return Arrays.asList(new ExactArgument("region"),
-                             new ExactArgument("flag"),
-                             ZONE_VALUE,
-                             new ExactArgument("tnt"),
-                             new ExactArgument("defuse"),
-                             new ExactArgument("set"),
-                             ENABLE_DISABLE);
+                new ExactArgument("flag"),
+                ZONE_VALUE,
+                new ExactArgument("damage"),
+                new ExactArgument("attack"),
+                new ExactArgument("towards"),
+                new ExactArgument("player"),
+                new ExactArgument("set"),
+                new ExactArgument("group"),
+                GROUP);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("Enable or disable the tnt defuse flag");
+        return Component.text("Sets the group for Entity Damage Player flag");
     }
 
     @Override
     public @NotNull Optional<ZonePermission> getPermissionNode() {
-        return Optional.of(ZonePermissions.FLAG_TNT_DEFUSE_ENABLE);
+        return Optional.of(ZonePermissions.FLAG_DAMAGE_ATTACK_SET_GROUP);
     }
 
     @Override
     public @NotNull CommandResult run(
             @NotNull CommandContext commandContext, @NotNull String... args) {
         Zone zone = commandContext.getArgument(this, ZONE_VALUE);
-        boolean enable = commandContext.getArgument(this, ENABLE_DISABLE);
-        TnTDefuseFlag tnTDefuseFlag = zone
-                .getFlag(FlagTypes.TNT_DEFUSE)
-                .orElse(new TnTDefuseFlag());
-        if (enable) {
-            zone.addFlag(tnTDefuseFlag);
-        } else {
-            zone.removeFlag(FlagTypes.TNT_DEFUSE);
-        }
+        EntityDamagePlayerFlag entityDamagePlayerFlag = zone
+                .getFlag(FlagTypes.ENTITY_DAMAGE_PLAYER)
+                .orElse(FlagTypes.ENTITY_DAMAGE_PLAYER.createCopyOfDefault());
+        Group group = commandContext.getArgument(this, GROUP);
+        zone.getMembers().addKey(group, entityDamagePlayerFlag.getRequiredKey());
         try {
             zone.save();
-            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.TNT_DEFUSE));
+            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.ENTITY_DAMAGE_PLAYER));
         } catch (ConfigurateException ce) {
-            ce.printStackTrace();
             return CommandResult.error(Messages.getZoneSavingError(ce));
         }
         return CommandResult.success();
