+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/player/attack/ZoneFlagDamageAttackView.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.damage.attack;
+package org.zone.commands.structure.region.flags.damage.to.player.attack;
 
 import net.kyori.adventure.identity.Identity;
 import net.kyori.adventure.text.Component;
