+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/player/fall/ZoneFlagPlayerFallDamageEnableDisable.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.damage.attack;
+package org.zone.commands.structure.region.flags.damage.to.player.fall;
 
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
@@ -7,27 +7,27 @@ import org.spongepowered.configurate.ConfigurateException;
 import org.zone.commands.system.ArgumentCommand;
 import org.zone.commands.system.CommandArgument;
 import org.zone.commands.system.arguments.operation.ExactArgument;
+import org.zone.commands.system.arguments.simple.BooleanArgument;
 import org.zone.commands.system.arguments.zone.ZoneArgument;
-import org.zone.commands.system.arguments.zone.ZoneGroupArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.permissions.ZonePermission;
 import org.zone.permissions.ZonePermissions;
 import org.zone.region.Zone;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.entity.player.damage.attack.EntityDamagePlayerFlag;
-import org.zone.region.group.Group;
+import org.zone.region.flag.entity.player.damage.fall.PlayerFallDamageFlag;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class ZoneFlagDamageAttackSetGroupCommand implements ArgumentCommand {
-
+public class ZoneFlagPlayerFallDamageEnableDisable implements ArgumentCommand {
     public static final ZoneArgument ZONE_VALUE = new ZoneArgument("zoneId",
             new ZoneArgument.ZoneArgumentPropertiesBuilder().setBypassSuggestionPermission(
-                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_ATTACK_SET_GROUP));
-    public static final ZoneGroupArgument GROUP = new ZoneGroupArgument("groupId", ZONE_VALUE);
+                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_FALL_ENABLE));
+    public static final BooleanArgument ENABLED = new BooleanArgument("enableValue",
+            "enable",
+            "disable");
 
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
@@ -35,37 +35,39 @@ public class ZoneFlagDamageAttackSetGroupCommand implements ArgumentCommand {
                 new ExactArgument("flag"),
                 ZONE_VALUE,
                 new ExactArgument("damage"),
-                new ExactArgument("attack"),
-                new ExactArgument("towards"),
-                new ExactArgument("player"),
+                new ExactArgument("fall"),
                 new ExactArgument("set"),
-                new ExactArgument("group"),
-                GROUP);
+                ENABLED);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("Sets the group for Entity Damage Player flag");
+        return Component.text("Command to enable/disable the fall damage flag");
     }
 
     @Override
     public @NotNull Optional<ZonePermission> getPermissionNode() {
-        return Optional.of(ZonePermissions.FLAG_DAMAGE_ATTACK_SET_GROUP);
+        return Optional.of(ZonePermissions.FLAG_DAMAGE_FALL_ENABLE);
     }
 
     @Override
     public @NotNull CommandResult run(
             @NotNull CommandContext commandContext, @NotNull String... args) {
+        boolean enable = commandContext.getArgument(this, ENABLED);
         Zone zone = commandContext.getArgument(this, ZONE_VALUE);
-        EntityDamagePlayerFlag entityDamagePlayerFlag = zone
-                .getFlag(FlagTypes.ENTITY_DAMAGE_PLAYER)
-                .orElse(FlagTypes.ENTITY_DAMAGE_PLAYER.createCopyOfDefault());
-        Group group = commandContext.getArgument(this, GROUP);
-        zone.getMembers().addKey(group, entityDamagePlayerFlag.getRequiredKey());
+        PlayerFallDamageFlag playerFallDamageFlag = zone
+                .getFlag(FlagTypes.PLAYER_FALL_DAMAGE)
+                .orElse(new PlayerFallDamageFlag());
+        if (enable) {
+            zone.addFlag(playerFallDamageFlag);
+        } else {
+            zone.removeFlag(FlagTypes.PLAYER_FALL_DAMAGE);
+        }
         try {
             zone.save();
-            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.ENTITY_DAMAGE_PLAYER));
+            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.PLAYER_FALL_DAMAGE));
         } catch (ConfigurateException ce) {
+            ce.printStackTrace();
             return CommandResult.error(Messages.getZoneSavingError(ce));
         }
         return CommandResult.success();
