+++ src/main/java/org/zone/commands/structure/region/flags/damage/to/player/fall/ZoneFlagPlayerFallDamageSetGroup.java
@@ -1,33 +1,33 @@
-package org.zone.commands.structure.region.flags.damage.fall;
+package org.zone.commands.structure.region.flags.damage.to.player.fall;
 
-import net.kyori.adventure.identity.Identity;
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.api.command.CommandResult;
+import org.spongepowered.configurate.ConfigurateException;
 import org.zone.commands.system.ArgumentCommand;
 import org.zone.commands.system.CommandArgument;
 import org.zone.commands.system.arguments.operation.ExactArgument;
-import org.zone.commands.system.arguments.operation.OptionalArgument;
 import org.zone.commands.system.arguments.zone.ZoneArgument;
+import org.zone.commands.system.arguments.zone.ZoneGroupArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.permissions.ZonePermission;
 import org.zone.permissions.ZonePermissions;
 import org.zone.region.Zone;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.group.key.GroupKeys;
+import org.zone.region.flag.entity.player.damage.fall.PlayerFallDamageFlag;
+import org.zone.region.group.Group;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class ZoneFlagPlayerFallDamageView implements ArgumentCommand {
+public class ZoneFlagPlayerFallDamageSetGroup implements ArgumentCommand {
 
     public static final ZoneArgument ZONE_VALUE = new ZoneArgument("zoneId",
             new ZoneArgument.ZoneArgumentPropertiesBuilder().setBypassSuggestionPermission(
-                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_FALL_VIEW));
-    public static final OptionalArgument<String> VIEW = new OptionalArgument<>(new ExactArgument(
-            "view"), (String) null);
+                    ZonePermissions.OVERRIDE_FLAG_DAMAGE_FALL_SET_GROUP));
+    public static final ZoneGroupArgument GROUP = new ZoneGroupArgument("groupId", ZONE_VALUE);
 
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
@@ -36,31 +36,36 @@ public class ZoneFlagPlayerFallDamageView implements ArgumentCommand {
                 ZONE_VALUE,
                 new ExactArgument("damage"),
                 new ExactArgument("fall"),
-                VIEW);
+                new ExactArgument("set"),
+                new ExactArgument("group"),
+                GROUP);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("View the details of Entity damage player flag");
+        return Component.text("Sets the group for Entity Damage Player flag");
     }
 
     @Override
     public @NotNull Optional<ZonePermission> getPermissionNode() {
-        return Optional.of(ZonePermissions.FLAG_DAMAGE_FALL_VIEW);
+        return Optional.of(ZonePermissions.FLAG_DAMAGE_FALL_SET_GROUP);
     }
 
     @Override
     public @NotNull CommandResult run(
             @NotNull CommandContext commandContext, @NotNull String... args) {
         Zone zone = commandContext.getArgument(this, ZONE_VALUE);
-        commandContext
-                .getCause()
-                .sendMessage(Identity.nil(),
-                        Messages.getEnabledInfo(zone.containsFlag(FlagTypes.PLAYER_FALL_DAMAGE)));
-        zone
-                .getMembers()
-                .getGroup(GroupKeys.PLAYER_FALL_DAMAGE)
-                .ifPresent(group -> commandContext.sendMessage(Messages.getGroupInfo(group)));
+        PlayerFallDamageFlag playerFallDamageFlag = zone
+                .getFlag(FlagTypes.PLAYER_FALL_DAMAGE)
+                .orElse(FlagTypes.PLAYER_FALL_DAMAGE.createCopyOfDefault());
+        Group group = commandContext.getArgument(this, GROUP);
+        zone.getMembers().addKey(group, playerFallDamageFlag.getRequiredKey());
+        try {
+            zone.save();
+            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.PLAYER_FALL_DAMAGE));
+        } catch (ConfigurateException ce) {
+            return CommandResult.error(Messages.getZoneSavingError(ce));
+        }
         return CommandResult.success();
     }
 }
