+++ src/main/java/org/zone/permissions/ZonePermissions.java
@@ -510,6 +510,67 @@ public enum ZonePermissions implements ZonePermission {
             "fire",
             "view"),
 
+    //zone visibility
+    ZONE_VISIBILITY_SET(false,
+            "zone",
+            "region",
+            "visibility",
+            "set"),
+    ZONE_VISIBILITY_VIEW(false,
+            "zone",
+            "region",
+            "visibility",
+            "view"),
+
+    //zone visibility override
+    OVERRIDE_ZONE_VISIBILITY_SET(false,
+            "zone",
+            "override",
+            "region",
+            "visibility",
+            "set"),
+    OVERRIDE_ZONE_VISIBILITY_VIEW(false,
+            "zone",
+            "override",
+            "region",
+            "visibility",
+            "view"),
+
+    //farmland trample
+    FLAG_FARMLAND_TRAMPLE_SET_ENABLE(false,
+            "zone",
+            "region",
+            "flag",
+            "farmland",
+            "trample",
+            "set",
+            "enable"),
+    FLAG_FARMLAND_TRAMPLE_VIEW(false,
+            "zone",
+            "region",
+            "flag",
+            "farmland",
+            "trample",
+            "view"),
+
+    //farmland trample override
+    OVERRIDE_FLAG_FARMLAND_TRAMPLE_SET_ENABLE(false,
+            "zone",
+            "override",
+            "region",
+            "flag",
+            "farmland",
+            "trample",
+            "set",
+            "enable"),
+    OVERRIDE_FLAG_FARMLAND_TRAMPLE_VIEW(false,
+            "zone",
+            "override",
+            "region",
+            "farmland",
+            "trample",
+            "view"),
+
     //balance
     FLAG_BALANCE_VIEW(true, "zone", "region", "flag", "balance", "view"),
