+++ src/main/java/org/zone/region/flag/FlagTypes.java
@@ -1,6 +1,8 @@
 package org.zone.region.flag;
 
 import org.jetbrains.annotations.NotNull;
+import org.zone.region.flag.entity.nonliving.block.farmland.FarmTramplingFlag;
+import org.zone.region.flag.entity.nonliving.block.farmland.FarmTramplingFlagType;
 import org.zone.region.flag.entity.nonliving.block.tnt.TnTDefuseFlagType;
 import org.zone.region.flag.entity.monster.move.PreventMonsterFlagType;
 import org.zone.region.flag.entity.player.damage.attack.EntityDamagePlayerFlagType;
@@ -16,8 +18,8 @@ import org.zone.region.flag.entity.player.move.preventing.PreventPlayersFlagType
 import org.zone.region.flag.meta.eco.balance.BalanceFlagType;
 import org.zone.region.flag.meta.eco.payment.buy.BuyFlagType;
 import org.zone.region.flag.meta.edit.EditingFlagType;
-import org.zone.region.flag.meta.join.request.JoinRequestFlagType;
-import org.zone.region.flag.meta.join.visibility.flag.ZoneVisibilityFlagType;
+import org.zone.region.flag.meta.request.join.JoinRequestFlagType;
+import org.zone.region.flag.meta.request.visibility.flag.ZoneVisibilityFlagType;
 import org.zone.region.flag.meta.member.MembersFlagType;
 import org.zone.region.flag.meta.tag.TagsFlagType;
 
@@ -33,6 +35,7 @@ public final class FlagTypes {
 
     public static final PreventMonsterFlagType PREVENT_MONSTER = new PreventMonsterFlagType();
     public static final ZoneVisibilityFlagType ZONE_VISIBILITY = new ZoneVisibilityFlagType();
+    public static final FarmTramplingFlagType FARM_TRAMPLING = new FarmTramplingFlagType();
     public static final JoinRequestFlagType JOIN_REQUEST = new JoinRequestFlagType();
     public static final MembersFlagType MEMBERS = new MembersFlagType();
     public static final DoorInteractionFlagType DOOR_INTERACTION = new DoorInteractionFlagType();
