+++ src/main/java/org/zone/region/flag/meta/request/visibility/flag/ZoneVisibilityFlag.java
@@ -1,10 +1,10 @@
-package org.zone.region.flag.meta.join.visibility.flag;
+package org.zone.region.flag.meta.request.visibility.flag;
 
 import org.jetbrains.annotations.NotNull;
 import org.zone.region.flag.Flag;
 import org.zone.region.flag.FlagType;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.meta.join.visibility.ZoneVisibility;
+import org.zone.region.flag.meta.request.visibility.ZoneVisibility;
 
 public class ZoneVisibilityFlag implements Flag {
